requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
voyageai==0.2.1
numpy==1.24.3